requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
python-multipart
python-dotenv
jinja2